DDP-7039 - remove unused fields; refactor classes structure
