+++ pepper-apis/src/main/java/org/broadinstitute/ddp/event/pubsubtask/api/PubSubTaskException.java
@@ -1,6 +1,5 @@
 package org.broadinstitute.ddp.event.pubsubtask.api;
 
-import com.google.gson.JsonObject;
 import org.broadinstitute.ddp.exception.DDPException;
 
 /**
@@ -16,7 +15,6 @@ public class PubSubTaskException extends DDPException {
     private final Severity severity;
     private final boolean shouldRetry;
     private final PubSubTask pubSubTask;
-    private final JsonObject extraPayload;
 
     public PubSubTaskException(String message, Throwable e) {
         this(message, e, null, null, false);
@@ -31,24 +29,14 @@ public class PubSubTaskException extends DDPException {
     }
 
     public PubSubTaskException(String message, Severity severity, boolean shouldRetry) {
-        this(message, null, severity, null, shouldRetry, null);
-    }
-
-    public PubSubTaskException(String message, Severity severity, PubSubTask pubSubTask, JsonObject extraPayload) {
-        this(message, null, severity, pubSubTask, false, extraPayload);
+        this(message, null, severity, null, shouldRetry);
     }
 
     public PubSubTaskException(String message, Throwable e, Severity severity, PubSubTask pubSubTask, boolean shouldRetry) {
-        this(message, null, severity, pubSubTask, shouldRetry, null);
-    }
-
-    public PubSubTaskException(String message, Throwable e, Severity severity, PubSubTask pubSubTask, boolean shouldRetry,
-                               JsonObject extraPayload) {
         super(message, e);
         this.severity = severity;
         this.pubSubTask = pubSubTask;
         this.shouldRetry = shouldRetry;
-        this.extraPayload = extraPayload;
     }
 
 
@@ -62,9 +50,5 @@ public class PubSubTaskException extends DDPException {
 
     public PubSubTask getPubSubTask() {
         return pubSubTask;
-    }
-
-    public JsonObject getExtraPayload() {
-        return extraPayload;
     }
 }
