--- pepper-apis/src/main/java/org/broadinstitute/ddp/event/pubsubtask/impl/userdelete/UserDeleteConstants.java
@@ -1,14 +1,0 @@
-package org.broadinstitute.ddp.event.pubsubtask.impl.userdelete;
-
-import org.broadinstitute.ddp.event.pubsubtask.api.PubSubTask;
-
-/**
- * Constants specific for taskType='USER_DELETE'
- */
-public class UserDeleteConstants {
-
-    /**
-     * {@link PubSubTask} taskType for user deletion
-     */
-    public static final String TASK_TYPE__USER_DELETE = "USER_DELETE";
-}
